// front-analysis-ia/src/app/models/Products.ts
export interface Producto{
  codigoProducto: string;
  nombreProducto: string;
  descripcion: string;
  especificaciones: string;
  precio: Number,
  existencias: Number,
  img: string;
  fechaCreacion: string;
  fechaModificacion: string;
  usuarioCreacion: string;
  usuarioModificacion: string;
  categoria: Number
}